Point cursor implementation
Added object and script to allow for point cursor to interact with the targets on the screen. Hovering over a target turns the target yellow, and on click, targets are destroyed. Also added Target.cs script to Target prefab in order to ensure that color changes correctly. In the future, still need to figure out how to change main target to red color.