Added functionality for making pie wheel appear
Made it so that the pie wheel can also be brought up on screen through pressing mouse button 5. This is done to make bringing up the pie wheel much more responsive than relying on pressing both left and right mouse at the exact same time. Left and right mouse button functionality is still left in place in case someone is using a mouse that does not have a side mouse/mouse 5 button.