Added text to each segment of pie wheel
Text has been added to show up on the pie wheel when it shows up on screen. It should be know that this is just 8 text objects that are being rendered over top of the pie wheel, rather than anything generated by code. However, this is a simple solution to the problem at hand, and can be something we go about making simpler through code in the future.